netdev::interface

Enum InterfaceType

pub enum InterfaceType {

    Unknown,
    Ethernet,
    TokenRing,
    Fddi,
    BasicIsdn,
    PrimaryIsdn,
    Ppp,
    Loopback,
    Ethernet3Megabit,
    Slip,
    Atm,
    GenericModem,
    FastEthernetT,
    Isdn,
    FastEthernetFx,
    Wireless80211,
    AsymmetricDsl,
    RateAdaptDsl,
    SymmetricDsl,
    VeryHighSpeedDsl,
    IPOverAtm,
    GigabitEthernet,
    Tunnel,
    MultiRateSymmetricDsl,
    HighPerformanceSerialBus,
    Wman,
    Wwanpp,
    Wwanpp2,
    Bridge,
    Can,

}

Type of Network Interface

Variants

Unknown

Unknown interface type

Ethernet

The network interface using an Ethernet connection

TokenRing

The network interface using a Token-Ring connection

Fddi

The network interface using a Fiber Distributed Data Interface (FDDI) connection

BasicIsdn

The network interface using a basic rate interface Integrated Services Digital Network (ISDN) connection

PrimaryIsdn

The network interface using a primary rate interface Integrated Services Digital Network (ISDN) connection

Ppp

The network interface using a Point-To-Point protocol (PPP) connection

Loopback

The loopback interface (often used for testing)

Ethernet3Megabit

The network interface using an Ethernet 3 megabit/second connection

Slip

The network interface using a Serial Line Internet Protocol (SLIP) connection

Atm

The network interface using asynchronous transfer mode (ATM) for data transmission

GenericModem

The network interface using a modem

FastEthernetT

The network interface using a Fast Ethernet connection over twisted pair and provides a data rate of 100 megabits per second (100BASE-T)

Isdn

The network interface using a connection configured for ISDN and the X.25 protocol.

FastEthernetFx

The network interface using a Fast Ethernet connection over optical fiber and provides a data rate of 100 megabits per second (100Base-FX)

Wireless80211

The network interface using a wireless LAN connection (IEEE 802.11)

AsymmetricDsl

The network interface using an Asymmetric Digital Subscriber Line (ADSL)

RateAdaptDsl

The network interface using a Rate Adaptive Digital Subscriber Line (RADSL)

SymmetricDsl

The network interface using a Symmetric Digital Subscriber Line (SDSL)

VeryHighSpeedDsl

The network interface using a Very High Data Rate Digital Subscriber Line (VDSL)

IPOverAtm

The network interface using the Internet Protocol (IP) in combination with asynchronous transfer mode (ATM) for data transmission

GigabitEthernet

The network interface using a gigabit Ethernet connection and provides a data rate of 1,000 megabits per second (1 gigabit per second)

Tunnel

The network interface using a tunnel connection

MultiRateSymmetricDsl

The network interface using a Multirate Digital Subscriber Line

HighPerformanceSerialBus

The network interface using a High Performance Serial Bus

Wman

The network interface using a mobile broadband interface for WiMax devices

Wwanpp

The network interface using a mobile broadband interface for GSM-based devices

Wwanpp2

The network interface using a mobile broadband interface for CDMA-based devices

Bridge

Transparent bridge interface

Can

Controller Area Network

Implementations

impl InterfaceType

pub fn value(&self) -> u32

Returns OS-specific value of InterfaceType

pub fn name(&self) -> String

Returns name of InterfaceType