netdev::interface

Struct Interface

pub struct Interface {

    pub index: u32,
    pub name: String,
    pub friendly_name: Option<String>,
    pub description: Option<String>,
    pub if_type: InterfaceType,
    pub mac_addr: Option<MacAddr>,
    pub ipv4: Vec<Ipv4Net>,
    pub ipv6: Vec<Ipv6Net>,
    pub flags: u32,
    pub transmit_speed: Option<u64>,
    pub receive_speed: Option<u64>,
    pub gateway: Option<NetworkDevice>,
    pub dns_servers: Vec<IpAddr>,
    pub default: bool,

}

Structure of Network Interface information

Fields
index: u32

Index of network interface

name: String

Name of network interface

friendly_name: Option<String>

Friendly Name of network interface

description: Option<String>

Description of the network interface

if_type: InterfaceType

Interface Type

mac_addr: Option<MacAddr>

MAC address of network interface

ipv4: Vec<Ipv4Net>

List of Ipv4Net for the network interface

ipv6: Vec<Ipv6Net>

List of Ipv6Net for the network interface

flags: u32

Flags for the network interface (OS Specific)

transmit_speed: Option<u64>

Speed in bits per second of the transmit for the network interface

receive_speed: Option<u64>

Speed in bits per second of the receive for the network interface

gateway: Option<NetworkDevice>

Default gateway for the network interface

dns_servers: Vec<IpAddr>

DNS servers for the network interface

default: bool

is default interface

Implementations

impl Interface

pub fn default() -> Result<Interface, String>

Construct a new default Interface instance

pub fn dummy() -> Interface

pub fn is_up(&self) -> bool

Check if the network interface is up

pub fn is_loopback(&self) -> bool

Check if the network interface is a Loopback interface

pub fn is_point_to_point(&self) -> bool

Check if the network interface is a Point-to-Point interface

pub fn is_multicast(&self) -> bool

Check if the network interface is a Multicast interface

pub fn is_broadcast(&self) -> bool

Check if the network interface is a Broadcast interface

pub fn is_tun(&self) -> bool

Check if the network interface is a TUN interface

pub fn is_running(&self) -> bool

Check if the network interface is running and ready to send/receive packets

pub fn is_physical(&self) -> bool

Check if the network interface is a physical interface